Remove ssh agent singleton
I really think this is not needed if I correctly name the key
files...what a waste of effort getting this to work in the first place.

This could be cleaned up to be silent and then used if I ever want to
start adding passwords to my key files.